Update and rename Circle bounce of each other to Code with laws of physics
I added damping factor when:
- the ball collides with each other;
- when ball collides with vertical walls;
- when the ball collides with bottom wall.
In physics, when they collide, some of energy is lost due to friction. So I make the dx and dy decrease by a certain value when it collides with another ball. When the ball collides with horizontal wall, only its x will decrease(because motions in different axis is separate). When ball collides with bottom wall, only dy decrease